Refactor MongoDB connection logic in db.js to use environment variable for URI, implement caching for connection, and remove unnecessary error handling comments.
