+++ config/db.js
@@ -1,90 +1,39 @@
+// db.js
+// MongoDB connection using Mongoose with best practices
+
 const mongoose = require('mongoose');
-const { secret } = require('./secret');
 
-mongoose.set('strictQuery', false);
+const MONGODB_URI = process.env.MONGO_URI;
 
-// local url
-const DB_URL = 'mongodb://0.0.0.0:27017/ewo';
-// mongodb url
-const MONGO_URI = secret.db_url;
+if (!MONGODB_URI) {
+  throw new Error(
+    'Please define the MONGODB_URI environment variable inside .env.local'
+  );
+}
 
-// const connectDB = async () => {
-//   try {
-//     await mongoose.connect(MONGO_URI);
-//     console.log('mongodb connection success!');
-//   } catch (err) {
-//     console.log('mongodb connection failed!', err.message);
-//   }
-// };
+let cached = global.mongoose;
 
-const connectDB = async () => {
-  try {
-    mongoose.set('strictQuery', false);
+if (!cached) {
+  cached = global.mongoose = { conn: null, promise: null };
+}
 
-    // Enhanced connection options
-    const connectOptions = {
-      useNewUrlParser: true,
-      useUnifiedTopology: true,
-      serverSelectionTimeoutMS: 15000, // 15 seconds
-      socketTimeoutMS: 45000, // 45 seconds
-      heartbeatFrequencyMS: 10000, // 10 seconds
-      retryWrites: true,
-      maxPoolSize: 10,
-      // Force node to use IPv4 over IPv6
-      family: 4,
-    };
-
-    const connection = await mongoose.connect(MONGO_URI, connectOptions);
-
-    console.log(
-      `MongoDB Connected: ${connection.connection.host.split('.')[0]}`
-    );
-
-    // Add global connection error handler to prevent app crashes
-    mongoose.connection.on('error', err => {
-      console.error('MongoDB connection error:', err);
-      // Don't throw the error - just log it to avoid crashing the application
-    });
-
-    // Handle disconnections and attempt to reconnect
-    mongoose.connection.on('disconnected', () => {
-      console.log('MongoDB disconnected. Attempting to reconnect...');
-      // Reconnection will be handled automatically by mongoose
-    });
-
-    return connection;
-  } catch (error) {
-    console.error('MongoDB connection failed:', error.message);
-
-    // Log but don't throw errors - prevent server crash on DB connection issues
-    // This allows the server to still handle API requests even if DB is down
-
-    // Optional: Set a reconnection timer
-    console.log('Will attempt to reconnect to MongoDB in 10 seconds');
-    setTimeout(() => {
-      console.log('Attempting MongoDB reconnection...');
-      dbConnect().catch(err => {
-        console.error('Reconnection attempt failed:', err.message);
+async function connectDB() {
+  if (cached.conn) {
+    return cached.conn;
+  }
+  if (!cached.promise) {
+    cached.promise = mongoose
+      .connect(MONGODB_URI, {
+        useNewUrlParser: true,
+        useUnifiedTopology: true,
+        // Add more options as needed
+      })
+      .then(mongoose => {
+        return mongoose;
       });
-    }, 10000);
   }
-};
-
-// Process-wide unhandled rejection handler to prevent crashes
-process.on('unhandledRejection', (reason, promise) => {
-  console.error('Unhandled Rejection at:', promise, 'reason:', reason);
-  // No need to exit process - let it continue running
-});
-
-// Process-wide uncaught exception handler to prevent crashes
-process.on('uncaughtException', error => {
-  console.error('Uncaught Exception:', error);
-  // Only exit on truly fatal errors
-  if (error.code === 'EADDRINUSE') {
-    console.error('Port in use, exiting...');
-    process.exit(1);
-  }
-  // For other errors, log but don't crash
-});
+  cached.conn = await cached.promise;
+  return cached.conn;
+}
 
 module.exports = connectDB;
